# api/openapi.yaml
openapi: 3.0.3
info:
  title: SentinelCore Management Hub API
  version: 1.0.0
  description: Zentrale API für Telemetrie, CIS-Hardening und hybrides Server-Provisioning über mTLS und WireGuard.
servers:
  - url: https://localhost:8443/api/v1
paths:
...
  /agent/v1/commands:
    get:
      summary: Nächsten tenant-gebundenen Befehl abholen
      security: [{ agentBearer: [] }]
      responses:
        '200': { description: Befehl geliefert }
        '204': { description: Keine ausstehenden Befehle }
...
components:
  securitySchemes:
    agentBearer:
      type: http
      scheme: bearer
      bearerFormat: shared-secret
...